+++ samples/data-stream-into.js
@@ -1,7 +1,7 @@
 #!/usr/bin/env node
 // module: data-stream, method: remap
 
-const DataStream = require('../').DataStream;
+const DataStream = require("../").DataStream;
 exports.log = console.log.bind(console);
 
 exports.test = async (test) => {
@@ -10,7 +10,7 @@ exports.test = async (test) => {
     const remapped = await DataStream.fromArray([[1,2,20], [3,4,21], [5,6,22], [7,8,23], [9,10,24]])
         .setOptions({maxParallel: 1})
         .into(async (out, item) => {
-            exports.log('it', item);
+            exports.log("it", item);
             await out.whenWrote(item[0]);
             return new Promise(s => process.nextTick(() => {
                 out.whenWrote(item[1]).then(s);
@@ -20,6 +20,6 @@ exports.test = async (test) => {
 
     exports.log("r", remapped);
 
-    test.deepEqual(remapped, [1,2,3,4,5,6,7,8,9,10], 'Emits items as programmed');
+    test.deepEqual(remapped, [1,2,3,4,5,6,7,8,9,10], "Emits items as programmed");
     test.done();
 };