+++ samples/data-stream-map.js
@@ -1,14 +1,14 @@
 #!/usr/bin/env node
 // module: data-stream, method: map
 
-const DataStream = require('../').DataStream;
+const DataStream = require("../").DataStream;
 
 exports.stream = () => DataStream.fromArray([1,2,3,4,5,6,7,8,9,10])
     .map(
         (num) => ({val: num})
     )
     .tap()
-    ;
+;
 
 // ------- END EXAMPLE --------