+++ scripts/tasks/index.js
@@ -1,0 +1,4 @@
+module.exports = {
+    tsd: require("./tsd"),
+    full_docs: require("./full-docs")
+};